+++ database/migrations/2016_08_13_145039_create_ingredients_table.php
@@ -19,7 +19,7 @@ class CreateIngredientsTable extends Migration
             $table->integer('id_unit') ->unsigned();
             $table->foreign('id_unit') ->references('id')->on('units');
             $table->string('ingrediente', 15);
-            $table->integer('stock');
+            $table->float('stock');
             $table->string('caracteristica', 40);
             $table->integer('stock_min');
             $table->timestamps();
@@ -36,4 +36,3 @@ class CreateIngredientsTable extends Migration
         Schema::drop('ingredients');
     }
 }
-
