+++ database/migrations/2016_08_13_150743_create_juices_has_ingredients_table.php
@@ -18,7 +18,7 @@ class CreateJuicesHasIngredientsTable extends Migration
             $table->foreign('juice_id')->references('id')->on('juices')->onDelete('Cascade');
             $table->integer('ingredient_id')->unsigned();
             $table->foreign('ingredient_id')->references('id')->on('ingredients')->onDelete('Cascade');
-            $table->integer('cantidad_ingrediente');
+            $table->float('cantidad_ingrediente');
             $table->integer('unit_id')->unsigned();
             $table->foreign('unit_id')->references('id')->on('units')->onDelete('Cascade');
             $table->timestamps();
